Dispatch baseband request to all registered local readout processes
Instead of sending a baseband request for each frequency processed on the
node (specified with the `freq_id` parameter of the request), the dump request
now applies to all frequencies on the node. The when the baseband request
manager receives the request, it adds it to the request queue of each registered
readout process (modifying the file name to include the process's frequency
index, and -- in the future -- the start time and length to take into account
the frequency-based delay).

This change implies the following changes to the format of the request:

1. the file name should not include the extension or the frequency index

2. element freq_id of the request is ignored, as the request now applies to all
   frequencies on the node

NB: when testing baseband dispatch, we now need a short wait (0.1 seems enough)
to allow processes to register with the basebandRequestManager before sending
baseband triggers.