Hot fix for critical DPDK alignment bug (#189)
There was a edge case where it was possible for DPDK to misalign FPAG packets by one or more frames (blocks of samples_per_data_set samples, or 49152 in the CHIME case). This of course causes de-correlation between cylinders.

Note this is a problem that can only happen at kotekan start up. So either it happens or it doesn't for a given run of a GPU instance. It's not something that can come and go. It also cannot cause small misalignments, so it should be easy to tell from the data when this has happened.

This fix simply kills kotekan if the alignment process fails on startup. We might want to add more code to try to realign things without killing kotekan, but as a hotfix, this seems to be the simplest and quickest way to ensure the problem doesn't impact the data going forward.