Jt/rfi sept2018 (#216)
Changes made for this pull request:

hsaRfiInputSum:

This kernel/command now computes and stores a mask for each SK estimate indicating if it should be removed. The config argument num_sigma determines the thresholds in unit of standard deviation of the SK estimator.

hsaRfiMaskOutput.cpp

This is a new copy-out kernel which copies the RFI mask to CPU. It also computes and atomic adds the total number of lost samples.

hsaRfiZeroData.cpp

A new kernel which zeros the appropriate parts of the input data. Kernel takes ~ 0.13% of the total compute time.

rfiBroadcast.cpp

Now consumes the RFI mask also. There is a new GET endpoint which returns the percent masked in the last frame.