Fix startup failure in HSA (the every other start failure) (#232)
Fixes a problem that caused kotekan to fail every other start. The problem showed up as a failure to process data fast enough. It turns out the function hsa_amd_agents_allow_access was being called every time a DMA memory transfer was queued. Every other run this function returned quickly, but in the other runs it would take anywhere from ~0.001 seconds to ~0.25 seconds every call. This was what was causing the system to fall behind, because delays of more than ~0.12 seconds total cannot be pipelined away.

This might related to which of the two hardware DMA queues is allocated to kotekan at startup.

The fix contained in this PR is call hsa_amd_agents_allow_access when the memory is allocated at startup.

Of interesting note, with this change calls to hsa_amd_agents_allow_access always take about the same (short) amount of time when only running at startup. So triggering this issue requires other items to be in the hardware queues.

Details of the change:

* Add new buffer type hsa which requires a GPU ID to set access too.
* Requires a GPU ID to be provided to hsa_host_malloc
* Removes calls to hsa_amd_agents_allow_access after startup
* Adds names to copy commands so they are correctly named in the profiling
* Updates configs to use the HSA buffers were needed.