Preallocate baseband dump buffer at readout process startup (closes #231)
Instead of allocating a new memory array for every baseband dump, this change
moves the allocation into the basebandReadout process's constructor, and just
reuses the segment for every dump. That way, the process should:

a) have the memory for the dump always ready, without having to go through an
allocation step;

b) prevent the allocator having to find a large buffer size even when the memory
is fragmented; and

c) fail right away if we are really allocating buffers that are too big for the
nodes' RAM.

The drawback of this change is that you can't start copying an event's data into
the baseband buffer while an earlier event is still being saved into the dump
file. A (future) fix could be to treat this space as a ring buffer or a similar
structure.