Preallocate baseband buffer (closes #231) (#245)
* Tweak baseband unit test wait times and max dump size

- The increased waits to startup should help with spurious test failures caused
  by not having enough time for the baseband readout processes to register.

- modified a test to properly test the handling of very long dumps (i.e., trim
  to configured max length)

- shortened the amount of time the pipeline has to wait for tests using advance
  start times, for quicker completion

* Preallocate baseband dump buffer at readout process startup (closes #231)

Instead of allocating a new memory array for every baseband dump, this change
moves the allocation into the basebandReadout process's constructor, and just
reuses the segment for every dump. That way, the process should:

a) have the memory for the dump always ready, without having to go through an
allocation step;

b) prevent the allocator having to find a large buffer size even when the memory
is fragmented; and

c) fail right away if we are really allocating buffers that are too big for the
nodes' RAM.

The drawback of this change is that you can't start copying an event's data into
the baseband buffer while an earlier event is still being saved into the dump
file. A (future) fix could be to treat this space as a ring buffer or a similar
structure.

* Coordinate baseband readout threads with condition variables

* Eliminate duplicate copies of `basebandDumpStatus`es

This change removes the second copy of basebandDumpStatus that was kept by the
basebandReadout as part of the communication between the listener and writer
threads, since the writer thread already gets a copy from the manager as the
result of the call to `get_next_ready_request`.

* Improve the API for basebandReadouManager::get_next_( waiting|ready )_request

Following Herb Sutter's guidelines for best practices around pointers/references/etc.:

1. use a unique_ptr to indicate a "maybe" result, with nullptr as "Nothing"
2. return a reference if the object can't be null

Therefore, instead out `get_next_waiting_request` returning a pair of pointers,
with "Nothing" indicated as nullptr members, just return a unique_ptr<pair>,
where the pair's members are *references*.

Similarly, `get_next_ready_request` also returns a pair of references, since
those members can never be null.