Abort on thread join failure
We have found a condition in which either the `basebandManager` and/or the `bufferSend` process fails to exit it's thread.  Both @cubranic and I have been trying to track it down but we haven't found it yet.  It might be deeper in the framework, and we are just seeing a symptom of it in those processes. 

In the meantime when this happens we should change from `exit()` to `abort()` since `exit()` doesn't always terminate the application in this failure mode.   This has been leaving some kotekan instances on the nodes in a "stuck" state where the binary doesn't auto restart, and so `kotekan_master` cannot get them running again.