Generic pattern test stage (#367)
This new visTestPattern is idle until an update is sent to it's endpoint. The update includes the test pattern FPGA buffer, a test name and number of frames. This stage will then run the given test (compare against the expected vis data calculated from the given FPGA buffer) for the given number of frames. Afterwards it is idle until the next update is received.
The endpoint is to be called by capo: https://github.com/chime-experiment/capo