Add dM endpoint: force register everything with the broker (#348)
This adds an endpoint called /dataset-manager/force-update. It can be used to force the dM to register everything it has with the broker. When the receiver node gets restarted, this can be used to force a GPU node to register stuff instead of doing so by restarting that node.

This also has unrelatedly changes the endpoint string in `restClient::make_request` to be passed by reference.