dataset management: hardcoded state type names (#419)
Adds hardcoded clean names for dataset types. Avoids issues with `typeid::name()` by instead mapping the `typeid::hash()` to a hardcoded name. 

* test(dataset_broker): save dset IDs on disk

This test calls 3 binaries that used to hardcode the dataset IDs.
Those dataset IDs had to be corrected every time the dataset code was
changed. To make that easier, the tests now store the dset IDs on disk
to share them between binaries.

* fix(datasetState): use hardcoded type names

Type names for serialization where created using RTTI. Those are not
guaranteed to be different for different types or to always be the
same for objects of the same type.

This adds a string to the REGISTER_STATE_TYPE macro and saves the
hardcoded name to a static map.

* fix(datasetManager): Use typeid::hash to compare types