Fix double counting of RFI and lost packets (#418)
It's possible that `hsaRfiMaskOutput.cpp` will count packet loss zeroed samples from `zeroSamples.cpp` twice.   This is because first `zeroSamples` adds to the lost samples counter for packet loss, and then later the RFI system might zero out the same samples (but in 256 time sample blocks), leading to double counting.   This fix checks if any of the samples zeroed in `zeroSamples` would be counted twice in `hsaRfiMaskOutput` and corrects for them. 

Also adds a new metadata value for number of RFI flagged samples.