Store counter metrics as 64-bit unsigned ints
Prometheus internally stores all metrics as doubles, but for us counters are non
only non-decreasing but also non-negative and incremented only by ints. So
64-bit unsigned int seems like the best fit in terms of native types -- there is
no need to do the incrementing in floating point, at least not until we change
the Counter API to allow fractional increments.

(As discussed on Slack with Andre.)