Move RFI metadata processing (#443)
Move RFI metadata processing out of the GPU stage into its own stage called `rfiUpdateMetadata`

This is trying to address the problem that when RFI flagging spikes, it requires checking the `lost_samples_buffer` for lost packets which happened at the same time as the RFI zeroed data.  This is to prevent double counting of RFI zeroed data, and lost packets.  However, this seemed to lead to instability when dealing with large RFI spikes, and caused a backlog in the GPU part of the system.