Update polyco script (#444)
We are approaching a time when the 26m will be operational again and we can resume testing/start observations with pulsar gating, so I improved the script that generates and sends gating configs. It should now be able to generate the config update and send it as well as disable gating all within the script.