Change the single dish VDIF reference epoch (#502)
This is not a real fix for #498. The change here will fail after the next leap second, and must be correctly fixed by #498. However until we fix that bug correctly, making the epoch for single dish VDIF match that of the CHIME beamformed data will allow the data to be aligned.