Double the size of the baseband data buffer
One drawback of BipBuffer is that it is supposed to be sized at least double the
expected (legal) reservation size. Otherwise, its read and write pointers can
get stuck halfway, at which point `write` can't advance because there is no
contiguous space in the buffer of the requested size, while `read` stays put
because there is no newly written data.

With buffer sized to even a byte over twice the size of maximum allowed dump
size, we know that if `read` has caught up to `write`, there will be available
space in the buffer for even the maximum dump.

Alternatively, we could use "reserve_max" for getting *any* write reservation,
even if smaller than requested. This might be more graceful when free buffer
space is running tight, as long as we're OK with writing dumps smaller than the
actually available sample data. (Note that in this case, the readout logic in
`basebandReadout::get_data` would also have to change slightly, since the
copying loop would have to exit when it's filled the reserved area, not at
`dump_end_frame`.)