fix: struct alignment calculation in python doesn't match kotekan
In numpy 1.16 the calculation of alignment for complex types was fixed
to match what happens in C. This exposed an issue in kotekan where we
are not padding the structs correctly. For the moment this has been
fixed by making numpy match kotekan (for all versions), though in the
long run we will probably change both back to C alignment.