refactor(visAccumulate): move frequency iteration, accumulate RFI, and moved accumulation state
A large change that heavily refactors the visAccumulate internals. This
moves the frequency iteration that is needed for the multifrequency
support, and the buffer frame management into the
initialise_output/finalise_output routines which substanially tidies up
the main loop.

Along with this, the accumulation state (fpga_seq_total/rfi_total) was
moved onto the individual datasets and the accumulation of the RFI
totals was added.