Allow empty element of `L1_node_ips` config param to denote inactive destination
Instead of having a REST API to completely disable sending to a beam's
destination, as removed in the preceding commit, this change introduces using an
empty string in the `L1_node_ips` list. Such nodes will not be checked for
liveness status and will always simply be skipped stream sending loop.