Revert the addition of CheckState
In the end, this state didn't add anything over the `live` flag, so I removed
it (but not reverted its commit entirely because it did add useful check
back-off logic). Note that one difference is that now both threads read the
value of the `live` atomic_bool flag, and I'm not sure if this might cause some
contention.