Redesign of baseband processing flow and synchronization (#519)
* Create a utility class for synchronized queues

The SynchronizedQueue class wraps `std::deque` and synchronizes access and
modification mutex using a mutex. All blocking callers can be interrupted by
calling the `cancel` method, after which the queue will not produce any more
results.

* Create BipBuffer data structure

BipBuffer, or bipartite buffer, is a single-producer single-consumer circular
buffer that provides writing a contiguous chunk of data. [^1]

This implementation is inspired by Andrea Lattuada and James Munns's description
of a lockless version that uses atomic read/write operations and careful
separation of reader/writer roles. [^2]

[^1]: https://www.codeproject.com/articles/3479/the-bip-buffer-the-circular-buffer-with-a-twist
[^2]: https://andrea.lattuada.me/blog/2019/the-design-and-implementation-of-a-lock-free-ring-buffer-with-contiguous-reservations.html

* Rework baseband request management (closes #497)

- Switch to using a BipBuffer for saving baseband data in `basebandReadout`.

- Use `SyncedQueue`s in the `basebandReadoutManager` to hand off requests to
  the readout and writeout threads, rather than the buggy and overly complicated
  system of pointers and condition variables used before.

- This also means that the readout and the writeout threads can operate
  independently from each other, rather than having exclusive access to the
  entire baseband data buffer. Therefore, the readout thread can start
  processing the next waiting request as soon as it's done copying the first
  request's samples into the readout buffer, and more than one request's data
  can be stored in the readout buffer, waiting to be written out to the disk by
  the writeout thread, as long as they fit.

* Rename basebandReadout methods to match thread terminology

In all discussions, we use terms "readout" and "writeout" thread, but the
methods that run those threads were for historical reasons named differently.
This change brings them in line with design documentation and practice.

* Distinguish in error messages between different ways reading out baseband can fail

This is principally used to tell the used in the baseband status when readout failed
because a write reservation was not available (i.e., the baseband buffer was
full), vs. when the trigger arrived too late, which used to be the only message
for failed readout.

* Double the size of the baseband data buffer

One drawback of BipBuffer is that it is supposed to be sized at least double the
expected (legal) reservation size. Otherwise, its read and write pointers can
get stuck halfway, at which point `write` can't advance because there is no
contiguous space in the buffer of the requested size, while `read` stays put
because there is no newly written data.

With buffer sized to even a byte over twice the size of maximum allowed dump
size, we know that if `read` has caught up to `write`, there will be available
space in the buffer for even the maximum dump.

Alternatively, we could use "reserve_max" for getting *any* write reservation,
even if smaller than requested. This might be more graceful when free buffer
space is running tight, as long as we're OK with writing dumps smaller than the
actually available sample data. (Note that in this case, the readout logic in
`basebandReadout::get_data` would also have to change slightly, since the
copying loop would have to exit when it's filled the reserved area, not at
`dump_end_frame`.)

* Remove xfail marker from the baseband tests

They should be reliable now, even on the CI server.

* Add a missing Doxygen file block

Fixes #497