Make rfi receiver script installable (#533)
* fix(rfi_receiver): remove unused imports
* fix(rfi_receiver): use python3 for version script
* feat(python): install rfi_receiver
* build(version.py): move into python kotekan module
* fix(rfi_receiver.py): use yaml.safe_load to suppress warning
* fix(rfi_receiver): don't use version.pty
* fix(rfi_receiver): use python logging module
* fix(rfi_receiver): encode str for socket write in python3
* fix(rfi_receiver): update hardcoded config
* fix(rfi_receiver): disable reading kotekan config

This was disabled because it ignores the structure
of the kotekan config, looks in the whole kotekan config tree for the
keys from the config hardcoded in this file. It ignores that the kotekan
config has variable names reused in different parts of the config and
just applies all of them, resulting in wrong values, i.e. for
samples_per_data_set.
To solve this issue, only the hardcoded config is used,
but it should become a config file or find a better way to read the
kotekan config.

* fix(rfi_receiver): header[num_local_freq] seems to be a list