Accelerate pinging of live destinations when they stop replying
Testing on the full cluster showed that pinging needs to handle the occasional
dropped response packet, since the core switch normally discards around 500
packets per second, and can go over 1,000 in bursts every few minutes.

The current scheme of doubling the ping interval up some configurable
limit (default 10 min), and marking the node dead when it hasn't responded for
more than the maximum ping interval means that if a single response is missed
when the node is being checked at the slowest ping frequency, it means the node
dead will be marked dead at its next scheduled check and then marked live again
within a few seconds when it's pinged again (since the reply usually gets
through fine).

Instead, we want to ping nodes at a regular, not too frequent pace, but switch
to several quick pings when a live node stops responding. If a reply is not
received within the `_ping_dead_threshold` (set to be sufficient for 5-6 quick
checks), the node is marked dead; if it is, the node stays live -- but either
way, after at most that amount of time has passed, the node will resume regular
scheduling.