fix(restClient): use bev_pair to pass requests into event thread
The restClient has a seperate thread that runs the libevent queue. New
rest requests where added from outside this thread. That lead to
sporadically lost requests and was not the way to do it according to
https://github.com/libevent/libevent/issues/733 .

To solve the issue, a bufferevent pair is added to pass new requests
into the event loop thread by pointer. There is a read-callback that
gets called inside the thread and schedules the new requests from inside.

Fixes #341