Handle dropped input frames in frbPostProcess
If a frame is dropped in any of the stage's inputs, this will be detected by
that input's current frame's starting fpga_seq_no being different from the other
inputs. When that happens, the stage will keep advancing the other inputs, until
all of them are at the same fpga_seq_no.