Move visFile to factory.hpp and skip unit tests for missing components (#585)
This change moves the visFile creation to use the standard factory.hpp
implementation. To do this some behaviour needed to be changes, in
particular some initialisation is now deferred until the first write
(rather than in the pseudo-constructor `create_file` method).

It also now skips tests that will fail because the required components are not built into kotekan. This required adding a new command line option `--version-json` to give back a machine parseable set of build options.

It also fixes a build error in `test_bip_buffer.cpp` on clang.