Fix RFI bad inputs update problems
Buffers updates to the bad input list, which is fed to the hsaRfiUpdateBadInputs command object.

Fixes a bug where two consecutive updates to the bad input list in <0.5s would result in the second update being lost. The bug was found as a result of investigating #594.

Also applies a fix for the race condition found in #600 which closes #594

Co-authored-by: andrerenard <andre@renard.io>
(cherry picked from commit b5fd669d89947a4a02bccfd4f942c736ac952d98)