fix(restClient): use bev_pair to pass requests into event thread
The restClient has a separate thread that runs the libevent queue. New
rest requests where added from outside this thread. That lead to
sporadically lost requests and was not the way to do it according to
libevent/libevent#733 .

To solve the issue, a bufferevent pair is added to pass new requests
into the event loop thread in an iovec structure. There is a read-callback that
gets called inside the thread and schedules the new requests from inside.

* use 20 MB watermark to make sure read callback can't starve threads dropping new requests
  This is to prevent the bev getting filled up so much that the read callback later starves other threads.
* style(restClient): merge log lines about status code and text

Re-enable tests:
if comet is installed, use that as the dataset broker in testing
use random ports everywhere (for parallel CI tests)
* re-enable test_dataset_broker
* re-enable test_dataset_manager_rest
* re-enable test_restclient
* re-enable test_freqslicer_broker:  needs to run serially, in order to not communicate with the same redis instance as test_dataset_broker
* ci(jenkins): run pytest with python3.7

some small fixes from playing around with cppcheck:
* style(kotekan): minimize scope of opt_val
* perf(kotekan): don't pass string.c_str() if not needed (nlohmann::json::parse accepts an std::string)
* style(config): don't copy expression for rethrowing
* style(config): remove unused variable
* fix(buffer.c): test pointer BEFORE accessing it not after
* style: remove return after exception in some places
* move revert_stack function to datasetState
* style(fakeVis): remove redundant cast

Fixes #341
Closes #255
Closes #341