Disable the L1 destination checking if `ping_dead_threshold` is 0
We are seeing recurring storms of nodes marking multiple destinations dead and
then live again within a few seconds, so it's clear either there is a bug in the
current implementation or the thresholds need to be fine-tuned. Since we are
pressed to start the run, it makes sense to add a way to disable this feature
from the configuration, should we want to do so without reverting it completely.