Check for baseband destination directory before reading out the samples (#640) (#647)
* Check for baseband destination directory before reading out the samples (#640)

Because `BasebandApiManager` is not a Stage, we don't have access to the running
configuration to construct the destination relative to `base_dir`, so the error
handling has to be done in the BasebandReadout. This is a little less ideal,
because this way we can't respond with an HTTP error as part of the request, but
still prevents using up the time to read out the samples that cannot be saved to
the requested destination.