Account for slow replies when deciding whether to ping again
This should handle the problem with the following scenario:

  1. send a ping at time T.

  2. because the target is live and has responded in the last (T - _ping_interval
    - _ping_dead_threshold) seconds, schedule the next ping on regular schedule
    at (T + _ping_interval)

  3. target reply is received at (T+1)

  4. at (T + _ping_interval), the target comes up for the next check, but because
    the most recent response was at less that (T - _ping_interval), i.e., at (T +
    1 - _ping_interval), we skip sending the ping and just schedule a regular
    check at (T + 2*_ping_interval)

  5. at (T + 2*_ping_interval), the target comes up for another check, and we send
    the ping. However, because it's been (2*_ping_interval - 1) since the
    last response, which is certainly greater than the _ping_dead_threshold, the
    target is marked dead until we receive its response

Instead, during a regular check, we send the ping as long as the last response
was received in (_ping_interval - _quick_ping_interval). We really don't need to
give that much slack, because the response would normally come back in less than
a second, but I didn't want to hard-code a constant value and using a slightly
more generous interval shouldn't cause any harm.