Introduce a startup phase during which checks are done more often
For the first "ping_interval", the nodes are pinged every "quick_ping_interval"
until they first respond. This should help to detect live nodes during startup,
so that a dropped response packet doesn't mean a 6-minute wait until the node is
checked again.

While there is a risk of a "thundering herd" with all nodes in the cluster
sending pings every which way, but since they are just starting, no L1 traffic
will be sent to destinations until they are detected as live.