Live ringmaps (#557)
This PR implements a live ring map maker for CHIME, that constantly generates ring maps from the incoming data and can serve them via an HTTP interface to requesting clients (as msgpack formatted data). These maps are the flux at each location in Jy (provided that data was similarly calibrated), and the visibilities can be apodized to reduce artefacts in the image domain.

Additionally this implements a Stage that will do a fully redundant stack on the incoming data stream, although eventually this should get merged into the `baselineCompression` stage.