feat(VisSharedMemWriter): new Stage and helpers to export visibility data.
An improvement upon visCalWriter. It exports the most recent samples of the
calibration data stream to a fixed-length ring-buffer-like shared memory region,
which can be ready by multiple external processes. This PR also provides a
Python class for efficiently reading the data from this buffer.

The shared memory region is composed of a Structured Data Region (summarises the
structure of the data), an Access Record region (records the writer's access to
the Data region), and the Data region (ring buffer which contains the data and
metadata from the frame, along with a valid byte).

It will also create a semaphore which will control access to the access record
of the shared memory region. The access record ensures the ability of readers to
inspect the data while the stage is continuously writing to the file. It will
communicate the validity of the frames. Frames are dropped if they arrive too
late or out of order.

This was the combined effort of Anja Boskovic, Rick Nitsche and Richard Shaw.

Pull request #686, closes #692, #704