feat(datasetManager): simplify dataset update
Also see https://github.com/chime-experiment/comet/pull/84

The endpoint /update-datasets does not require nor return any list of
known roots nor timestamps anymore. For this, the update_datasets()
method was simplified here.

It also doesn't return anything but the datasets between the given one
and its root dataset anymore. It doesn't base the set of returned
datasets on any timestamp, but always returns the whole branch all the
way to the root. This does not require any changes in the datsetManager.