fix(shared_memory_buffer): upon update, first check if link still exists to shared_mem
If kotekan-calbuffer is restarted, then it will write the buffer to a
new location in shared memory, but the calibration broker will continue
to check old location, and will continue to see the same old data, which
it will ignore.

The SharedMemoryReader should check on every .update() call to
see if the file descriptor still has any links.

mmap does not contain a reference to the file descriptor used to create
it, and so we will need to keep the original file descriptor around.

Addresses https://github.com/chime-experiment/ch_cal/issues/2