initial push for freq-encoding changes in pulsar packet (#821)
This is the first in likely a series of changes to the pulsar post-processing code as we gradually move away from the VDIF structure. for starters, the "major" change i've made is to use previously-untouched parts of the current structure to encode frequency information, so that we can move away from the external map-retrieval process we've had to use since june 2020.

for now we technically still use the old VDIFHeader structure definition but truly revamp to remove irrelevant features / add relevant feature once the key pulsar folks have/make time for converging on a plan.