Revert "Increase maximum baseband dump duration to 1 s (#832)"
The reason the maximum duration was reduced from 1s to 500 ms in the first
place, was because `basebandReadout` allocates the readout buffer slightly more
than twice as long as the maximum dump duraton so it can't get stuck halfway in
BipBuffer. This uses too much memory with 1s maximum duration, unfortunately.

This reverts commit 16615f7dac3a62c282ec99bda612a094229cbecf.