Revert blaze version to v3.4 (#851)
When running on site with blaze version `v3.8-12-gf36b481f2` we discovered a regression in the performance of the eigenvalue decomposition.   This change reverts to the `v3.4` API and locks the version in the CI server. 

Also locks `h5py` to `2.9.0` since that broke while I was doing this PR.