feat(RfiFrameDrop): register a dataset state
- add RfiFrameDropState
  - enabled or not
  - thresholds
- register it in RfiFrameDrop stage

The state is where the thresholds and fractions are kept now in order to
avoid duplicates.
The state changes together with it's ID whenever *any* of the two
endpoint callbacks is executed. This is why all those values are
protected with a mutex now instead of by handing over shared pointers.