Remove redundant baseband status locking (#887)
Baseband API manager needs to acquire a lock on the request currently processed
by either the readout or the writeout thread, in order to see it in a consistent
state. For the endpoint returning only a single event's status, the manager does
not need to acquire this lock if the readout/writeout threads are not processing
this particular event.