Remove the check for the existence of the baseband destination path
This uses the `stat()` system call and takes a _very_ long time when the
destination NFS server is heavily loaded (such as when the full-cluster baseband
write for an earlier event is still going on).

Instead, we can just read out the data into the baseband buffer, which takes
only a second, and fail in the writeout thread when it tries to create the
destination file.