Break out `basebandReadout::get_data` into `wait_for_data` and `extract_data`
`get_data()` was a long function, with multiple nested loops, which made it
difficult to read. Logically, it performed two different functions: wait until
all the data is in the ring buffer, and then copy it to the baseband
buffer (or the output buffer, in the new design). Therefore, it's a
straightforward refactoring to just break it in two.