Pack requested event data across output frames
This commit builds on the previous one, but with no empty space in output frames
except in the very last one: as we finish copying event data from an input
frame, we don't flush the output until it's full, and just continue filling it
with the next input frame.

The code still expects the input and output frame sizes to be the same, but
that's not a limitation of the algorithm, just convenience using what's already
in the config and class fields.