Send baseband to a receiver node instead of writing files directly (#927)
This change tears out the logic that existed to write baseband into files (i.e., over NFS) from the basebandReadout stage, and makes it truly just a "readout": when a baseband request arrives, the stage now copies out the requested data into an output Kotekan buffer, with event details recorded in the output frame metadata. The idea is that the output buffer will simply be forwarded to a baseband receiver node using a `bufferSend` stage.  Also adds multi-frequency support.

Co-authored-by: chitrang <chitrang.patel@mail.mcgill.ca>
Co-authored-by: Andre Renard <andre@renard.io>