Make clang-format work with version 8 and 14 (#1075)
Ideally we'd bump the clang-format version to 14 or higher, but due to the current CI system being locked to ubuntu 18.04, we cannot do that without also making lots of other changes. This change appears to generate a state that's compatible with both version 8 and 14.   We should bump the ubuntu version in the CI system, but this will be done in another PR.