Add support for more than 3 cuda streams and custom synchronization between them (#1076)
Currently there are 3 cuda streams, two for copies, and one for compute.  This change allows for an arbitrary number of streams set by the `num_streams` option.   When using this option it becomes necessary to  provide a `cuda_stream` for each commandObject. 

To synchronize between streams, a new command object called `cudaSyncStream` can be used to place cuda wait events which depend on other streams.  There are also simple objects for syncing the compute and data copy streams for the simple base cases. 

All configurations moving forward will need to use this object to sync between the kernel and copy streams. 

---------

Co-authored-by: Dustin Lang <dstndstn@gmail.com>