Fix a memory lock issue with restricted docker instances. (#1079)
In #1074 we introduced an issue for restricted docker containers which don't allow memory locking or adjusting the memory policy.  This was preventing some github CI tests from running.   To fix this we add back the `NO_MEMLOCK` compiler option which removes the binary's ability to make memory policy changes.

This change also introduces a "zero_new_frames" config option to override the default behavior of zeroing new frames.   This is useful in situations where a stage doesn't assume anything about the frame contents (true for almost all stages), and so can reduce the startup time by just not zeroing the memory.   The option is `true` by default to preserve backwards compatibility with any configs that assume it's true. 

---------

Co-authored-by: Dustin Lang <dstndstn@gmail.com>