Fix Release build and add to CI (#1315)
Small changes to make release builds compile and run, largely by annotating variables and parameters which were only used in asserts or other debug testing as `[[maybe_unused]]`.  Adds a gcc release build with all tests to CI, as well as a clang release build (tests are not run).  Added `-Wno-unused-variable` to the global CMake in release to suppress errors about variables in auto-generated code.  A future PR should alter the generator to annotate the problem variables so the check can be restored.